Move drive util to util folder
