+++ src/main/java/frc/robot/subsystems/Drive.java
@@ -23,10 +23,10 @@ import edu.wpi.first.math.kinematics.DifferentialDriveOdometry;
 import edu.wpi.first.math.kinematics.DifferentialDriveWheelSpeeds;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
-import frc.robot.driveutil.DriveConfiguration;
-import frc.robot.driveutil.TJDriveModule;
 import frc.robot.util.SyncPIDController;
 import frc.robot.util.WrappingPIDController;
+import frc.robot.util.drive.DriveConfiguration;
+import frc.robot.util.drive.TJDriveModule;
 import frc.robot.util.preferenceconstants.DoublePreferenceConstant;
 import frc.robot.util.preferenceconstants.PIDPreferenceConstants;
 import frc.robot.util.transmission.CTREMagEncoder;
