+++ src/main/java/frc/robot/util/drive/DriveUtils.java
@@ -1,4 +1,4 @@
-package frc.robot.driveutil;
+package frc.robot.util.drive;
 
 import java.lang.Math;
 import java.util.function.DoubleSupplier;
